requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
openai>=1.40.0
python-docx>=1.1.0
pypdf>=4.2.0
chardet>=5.2.0
